Move load config logic to schutil package
Add LoadConfigFromFile function to schutil package.
Create stripPkgsArrayFromBody function to abstract "pkgs" parsing logic
out from lambda handler.
Run gofmt on all codebase.
