+++ src/main.go
@@ -1,18 +1,14 @@
 package main
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"math/rand"
 	"net/http"
-	"net/http/httputil"
-	"net/url"
 	"os"
 	"regexp"
-	"strconv"
 	"strings"
 	"time"
 
@@ -21,45 +17,47 @@ import (
 	"github.com/urfave/cli"
 )
 
-type Config struct {
-	Host          string   `json:"host"`
-	Port          int      `json:"port"`
-	Balancer      string   `json:"balancer"`
-	LoadThreshold int      `json:"load-threshold"`
-	Workers       []string `json:"workers"`
-}
-
 // Global variables
 var workers []schutil.Worker
-var config Config
+var config schutil.Config
 var supportedBalancers = []string{"random", "pkg-aware", "round-robin", "least-loaded"} // TODO(Gus): Change to a map[string]func
 
-func DoRunLambda(w http.ResponseWriter, r *http.Request) (*schutil.Worker, *schutil.HttpError) {
-	body, _ := ioutil.ReadAll(r.Body)
-	strBody := string(body)
+func stripPkgsArrayFromBody(strBody string) ([]string, string, *schutil.HttpError) {
 	pkgsRegExp := regexp.MustCompile(`"pkgs"\s*:\s*\[.*\],*\s*`)
 	matches := pkgsRegExp.FindStringSubmatch(strBody)
 	if len(matches) < 1 {
-		return nil, &schutil.HttpError{"Pkgs array required", http.StatusInternalServerError}
+		return nil, "", &schutil.HttpError{"Pkgs array required", http.StatusInternalServerError}
 	}
 	strPkgsJson := matches[0]
 	pkgsArrayRegExp := regexp.MustCompile(`\[.*\]`)
 	srtPkgsMatches := pkgsArrayRegExp.FindStringSubmatch(strPkgsJson)
 	if len(srtPkgsMatches) < 1 {
-		return nil, &schutil.HttpError{"Pkgs array ill-formed", http.StatusInternalServerError}
+		return nil, "", &schutil.HttpError{"Pkgs array ill-formed", http.StatusInternalServerError}
 	}
 	decoder := json.NewDecoder(strings.NewReader(srtPkgsMatches[0]))
 	var pkgs []string // pkgs ordered from larger to smaller
 	err := decoder.Decode(&pkgs)
 	if err != nil {
-		return nil, &schutil.HttpError{err.Error(), http.StatusInternalServerError}
+		return nil, "", &schutil.HttpError{err.Error(), http.StatusInternalServerError}
 	}
 
-	{ // Modify request's body
-		newStrBody := strings.Replace(strBody, strPkgsJson, "", -1)
-		r.Body = ioutil.NopCloser(strings.NewReader(newStrBody))
-		r.ContentLength = int64(len(newStrBody))
+	newStrBody := strings.Replace(strBody, strPkgsJson, "", -1)
+
+	return pkgs, newStrBody, nil
+}
+
+func DoRunLambda(w http.ResponseWriter, r *http.Request) (*schutil.Worker, *schutil.HttpError) {
+	body, _ := ioutil.ReadAll(r.Body)
+	strBody := string(body)
+	pkgs, newStrBody, err := stripPkgsArrayFromBody(strBody)
+
+	if err != nil {
+		return nil, err
 	}
+
+	// Modify request's body
+	r.Body = ioutil.NopCloser(strings.NewReader(newStrBody))
+	r.ContentLength = int64(len(newStrBody))
 
 	{ // Select worker and serve http
 		var worker *schutil.Worker
@@ -122,41 +120,16 @@ func StatusHandler(w http.ResponseWriter, r *http.Request) {
 
 func StartServer(ctx *cli.Context) {
 	configFilepath := ctx.String("config")
-	file, rfErr := ioutil.ReadFile(configFilepath)
-	if rfErr != nil {
-		log.Fatalf("Cannot read config file (%s)", configFilepath)
-	}
-	decoder := json.NewDecoder(bytes.NewReader(file))
-	jsonErr := decoder.Decode(&config) // Parse json config file
-	if jsonErr != nil {
-		log.Fatalf("Config file Ill-formed (%s)", configFilepath)
-	}
+	config = schutil.LoadConfigFromFile(configFilepath)
+	workers = schutil.CreateWorkersArray(configFilepath, config)
 
-	if len(config.Workers)%2 != 0 {
-		log.Fatalf("Config file Ill-formed (%s), every worker must have a weight", configFilepath)
-	}
-
-	for i := 0; i < len(config.Workers); i = i + 2 { // Make Workers with their Reverse Proxy Handlers
-		weight, err := strconv.Atoi(config.Workers[i+1])
-		if err != nil || weight < 0 {
-			log.Fatalf("Config file Ill-formed (%s), every worker weight must be a positive number", configFilepath)
-		}
-		u, _ := url.Parse("http://" + config.Workers[i])
-		handler := httputil.NewSingleHostReverseProxy(u)
-		workers = append(workers, schutil.Worker{u, handler, 0, weight})
-	}
-
-	// http.HandleFunc("/scheduler", func(w http.ResponseWriter, r *http.Request) {
-	//     fmt.Fprint(w, "I am a humble scheduler")
-	// })
 	http.HandleFunc("/runLambda/", RunLambdaHandler)
 	http.HandleFunc("/status", StatusHandler)
-	rand.Seed(time.Now().Unix()) // For rand future calls
 	log.Print("Scheduler is running")
 	http.ListenAndServe(fmt.Sprintf("%s:%d", config.Host, config.Port), nil)
 }
 
-func main() {
+func createCliApp() *cli.App {
 	app := cli.NewApp()
 	app.Usage = "Scheduler for Open-Lambda"
 	app.UsageText = "olscheduler COMMAND [ARG...]"
@@ -179,5 +152,12 @@ func main() {
 			Action: StartServer,
 		},
 	}
+	return app
+}
+
+func main() {
+	rand.Seed(time.Now().Unix()) // For rand future calls
+
+	app := createCliApp()
 	app.Run(os.Args)
 }
